Extract chat nudge state
