+++ dev-docs/architecture.md
@@ -65,6 +65,7 @@ js/
   chat-window-bindings.js — Chat callback wiring and legacy window exports
   chat-marker-prompts.js — Per-marker and selected-correlation AI prompt builders
   chat-empty-state.js — Empty chat and onboarding message renderer
+  chat-nudge.js      — Chat FAB nudge badge and dismissal state
   chat-threads.js    — Chat thread index CRUD and rail rendering
   chat-thread-search.js — Thread rail message search and match highlighting
   client-list.js    — Client List modal: search/sort/filter profiles, inline CRUD, archive/flag/pin
