+++ js/chat.js
@@ -22,9 +22,10 @@ export {
 } from './chat-actions.js';
 export {
   closeChatPanel, getChatWebSearchEnabled, refreshWebSearchToggle,
-  setChatNudge, setChatWebSearchEnabled, toggleChatFullscreen,
-  toggleChatPanel, openChatPanel, updateChatInputState, updateChatNudge,
+  setChatWebSearchEnabled, toggleChatFullscreen,
+  toggleChatPanel, openChatPanel, updateChatInputState,
 } from './chat-panel.js';
+export { setChatNudge, updateChatNudge } from './chat-nudge.js';
 export {
   cleanupDiscussionState, continueDiscussion, endDiscussion,
   getCurrentDiscussionState, getThreadPersonaCount,
