+++ service-worker.js
@@ -75,6 +75,7 @@ const APP_SHELL = [
   '/js/chat-marker-prompts.js',
   '/js/chat-attestation.js',
   '/js/chat-actions.js',
+  '/js/chat-nudge.js',
   '/js/chat-panel.js',
   '/js/chat-discussion.js',
   '/js/chat-onboarding.js',
